Rename "is" as "is-js"
This is just to make the same code work properly on both Node and frontend
side. There's a nasty naming inconsistency so I had to perform a little hack
(bound window.is as window['is-js']). Fortunately it seems to work.

Ideally the author of is-js will change the naming scheme. Alternatively I
could maintain a version of my own. There are some other changes that would be
beneficial as well.